Fix problems in handling of CLI "exec" mode and "config" mode commands.
Root cause analysis:

Commands with a "prompt/answer" sequence require
the user to specify a dictionary containing the command, prompt,
and prompt "answer". The previously existing handling enclosed the
entire (quoted) command entry under a single "command" dictionary
item for the dictionary that is sent to the switch instead of
setting the required specified values for each of the three
dictionary items.

Fix:
 - Create a utility function to generate the correct dictionary
   entries from the input command list parsed from the user
   playbook.
 - Use the resulting dictionary instead of the "raw" list
   of quoted command strings provided by the "AnsibleModule"
   parsing invocation.