Fix problems in handling of CLI "exec" mode and "config" mode commands.
Root cause analysis:

Commands containing  a "prompt/answer" sequence require
the user to specify a dictionary containing the command, prompt,
and "answer" (for the prompt). The previously existing handling enclosed the
entire command dictionary (containing the "command", "prompt",
and "answer" key/value pairs) in quotes as the value for an
unintended outer "command" key. The other items in the
dictionary input to the "send_command" API (for sending
commands to a device) were set to defaults (for exec commands)
or left unspecified (for configuration commands).

For example:

- Ansible "module" contents from parsing a playbook that specified
an exec command "list" containing a single command/prompt/answer sequence:

-----------------------------
["{'command': 'image remove all', 'prompt': '\\\\[y/N\\\\]:', 'answer': 'N'}"]
-----------------------------

- Incorrect result of formatting the "raw" single command list shown
above to prepare for invoking the "send_command" API:

-----------------------------
{'command': "{'command': 'image remove all', 'prompt': '\\\\[y/N\\\\]:', 'answer': 'N'}", 'output': None, 'prompt': None, 'answer': None, 'newline': True, 'sendonly': False, 'check_all': False}
-----------------------------

- Instead of the incorrect dictionary shown above, the desired dictionary
is:
-----------------------------
{'command': 'image remove all', 'prompt': '\\\\[y/N\\\\]:', 'answer': 'N', 'newline': True, 'sendonly': False, 'check_all': False}
-----------------------------

Fix:

    - Create a utility function to generate the correct dictionary
    entries from the input command list parsed from the user
    playbook.
    - Use the resulting dictionary instead of the "raw" list
    of quoted command strings provided by the "AnsibleModule"
    parsing invocation to formulate the correct arguments
    for the "send_command" API.