Fix the regression test failure for the aaa resource module.
This change modifies the testing of the "delete" test case for the
aaa resource module to enable testing of the functionality without
causing a failure in the "idempotent" test for the subsequent
merge/restore test case.

Problem analysis:

The "delete" test (Delete aaa properties; test_case_04) is
intended to verify functionality for deletion of an aaa configuration
attribute. Deletion of the "fail_through" bool attribute, however,
has the side effect in the SONiC functional code of also deleting
the configuration for the other aaa configuration attributes
("group" and "local"). Changing it from the "null" (deleted)
state back to a bool value also deletes the other attributes.
Because of this side effect when "restoring" the "fail_through"
attribute, the merge/restore test (aaa properties; test_case_05) fails
during the "idempotent" section of the test because the specified
values for the "group" and "local" attributes do not remain as
specified when simultaneously restoring the deleted value of the
of the "fail_through" attribute.

Fix:

Test/verify the "delete" functionality using an attribute
that does not have side effects on the other attributes:
The "group" attribute is a suitable choice for this testing.