Create release 1.1.1 changelog and other official Ansible release artifacts
This changeset contains the archive source fragments used for generating
the Ansible release changelog, the generated changelog, and other
files generated by the antsibull-changelog tool for the release.

The galaxy.yml file and the changelogs/config.yaml file have also
been updated to specify the new galaxy version information, to add
new authors for the Dell SONiC Enterprise Ansible collection and to save
changelog fragments in an archive folder.