Change the "prefix_list" resource module name to "prefix_lists".
This change is being made to conform to the naming convention used for other
resource module names in the Dell SONiC Enterprise Ansible collection
and also in the Ansible collections for most other vendors.
For resources that can have an arbitrary number of objects of the same
type, the plural form of the resource type is used.