Template for Pull Request
This is PR is an extension of [PR#121](https://github.com/ansible-collections/dellemc.enterprise_sonic/pull/121) and [PR#139](https://github.com/ansible-collections/dellemc.enterprise_sonic/pull/139). We have added all the comments from PR121 to this PR and have commented it below.

@kerry-meyer Thanks for creating these guideline. I have posted some minor comments/questions inline.

kerry-meyer In the PR description, please provide any relevant details on recommended procedures and/or tools for measuring code coverage (if any).

KVSK We have commented the below statement under additional information section to handle this code coverage measurement.
"Measure the code coverage before and after the change by running the UT and ensure that the "coverage after the change" is not less than the coverage "before the change". Note that the unit testing coverage can be manually executed using the pytest tool or ansible-test tool."

kerry-meyer Please change this to:
I have maintained backward compatibility or have provided any relevant "breaking_changes" descriptions in a "fragment" file in the "changelogs/fragments" directory of this repository.

KVSK Addressed as suggested.

kerry-meyer What does "vetting" mean in this requirement? Is there a recommended procedure to check for "security issues" (or is this a judgement call for the PR submitter)?

KVSK Have requested the technical team to comment on this.

kerry-meyer Please add an item for documentation requirements:
I have provided a summary for this PR in valid "fragment" file format in the "changelogs/fragments" directory of this repository branch. (REF: "https://docs.ansible.com/ansible/devel/community/development_process.html#changelogs-how-to")

KVSK Addressed as suggested.

kerry-meyer Please add a sentence here:
If no Github issues are affected, please indicate this with "N/A".

KVSK Addressed as suggested.