Fix sanity errors caused by deletion of Ansible Netcommon utilities (#172)
* Fix sanity errors caused by deletion of Ansible Netcommon utility functions.

Ansible recently merged changes in the Ansible Netcommon repo that eliminated the
source functions for several utilities imported into modules in the enterprise_sonic
collection. The affected utility functions are available in the ansible/ansible
(library) repo and have been there since the earliest Ansible core version
that is currently supported. Because of this, the source of the imports
can be changed in the enterprise_sonic collection without introducing any
downward compatibility problems.

* Deleted an unintentional extra line from the change set.