Signed-off-by: Kerry Meyer <kerry.meyer@dell.com>
Change the "galaxy" dependency statement for Ansible Netcommon to allow installation of any release after 5.0.0.

- Remove the temporary workaround merged to avoid installation of the bug in Netcommon version 5.0.0: Allow
installation of any Netcommon release after 5.0.0. This allows installation of the latest tagged release for
Ansible Netcommon, which contains the fix for the problem created with the 5.0.0 release plus additional
updates.

- Update the enterprise_sonic release number in preparation for releasing a minor version of the enterprise_sonic
collection that addresses backing out of a "breaking" fix in the AAA module and also the removal of the temporary
workaround for the Ansible Netcommon bug.