Galaxy yml fix (#270)
* Signed-off-by: Kerry Meyer <kerry.meyer@dell.com>

Change the "galaxy" dependency statement for Ansible Netcommon to allow installation of any release after 5.0.0.

- Remove the temporary workaround merged to avoid installation of the bug in Netcommon version 5.0.0: Allow
installation of any Netcommon release after 5.0.0. This allows installation of the latest tagged release for
Ansible Netcommon, which contains the fix for the problem created with the 5.0.0 release plus additional
updates.

- Update the enterprise_sonic release number in preparation for releasing a minor version of the enterprise_sonic
collection that addresses backing out of a "breaking" fix in the AAA module and also the removal of the temporary
workaround for the Ansible Netcommon bug.

* Signed-off-by: Kerry Meyer <kerry.meyer@dell.com>

Change the enterprise_sonic version number to 2.2.0

In accordance with a request from Ansible for corrected semantic versioning associated
with "breaking" changes and backing out of breaking changes, the "MINOR" version number
for this "fix" release is being incremented instead of just the "PATCH" version.

* Signed-off-by: Kerry Meyer <kerry.meyer@dell.com>

Add a fragment file for updating the galaxy.yml file and update other changelog files.

The additional changes beyond adding the new fragment file are to move "leftover"
fragment files that were not incorporated into the 2.1.0 release notes into the
changelogs/archive_fragments/2.1.0 directory and add the corresponding updates
to the changelogs/changelog.yaml file.