Update README.md
Update the "README.md" file to recommend an ansible-core version of 2.17 or higher and a Python version of 3.10 or higher. This matches the current support status specified in the ansible-core-support-matrix ( https://docs.ansible.com/ansible/latest/reference_appendices/release_and_maintenance.html#ansible-core-support-matrix ) and is required for the latest release of the enterprise_sonic Ansible resource module collection.